backend/requirements-dev.txt
# Developer tooling for Python backend
# Installed into the local venv at backend/.venv

# Formatters / Linters
black
ruff
isort
flake8

# Static analysis / complexity
radon
mypy
pylint
bandit
safety

# Security/code scanning CLI
semgrep

# Testing coverage (Week 4)
greenlet==3.2.4  # Required by SQLAlchemy async engine for tests
pytest-cov
pytest-timeout

